Separate DBManager into DBManager and DBClient

+++ refactoring/src/main/java/ru/akirakozov/sd/refactoring/Product.java
@@ -1,3 +1,3 @@
 package ru.akirakozov.sd.refactoring;
 
-public record Product(String name, int price) {}
+public record Product(String name, long price) {}
